Install StyleCop.Analyzers 1.0.0-alpha005
Some rules are disabled/hidden due to not being quite to (legacy) StyleCop's standards or that our code base is not up to par. The goal is to retain our 0 warnings build, while adding stylecop to keep our code base at its current quality, and over time enable rules and improve it.