Fix Report class to actually assert when failures occur
Before this, Report was itself calling Debug, which compiles away to a no-op in Release builds. So as a result, MS.VS.Validation.dll used by anyone else never ever popped debug assert dialogs, contrary to its original design.
With this change, we use the Trace class which skips the "DEBUG" symbol check in the calling code so that it always occurs. We add our own DEBUG check to the Report class's methods so that it has the originally intended effect of only validating on Debug builds of the calling code.