Overload Requires.NotNullOrEmpty for ICollection<T>
Previously, calling `Requires.NotNullOrEmpty<T>` with an `IEnumerable<T>` type was guaranteed to allocate an enumerator, even if the concrete type had a non-allocating `Count` property, such as `ICollection<T>`.

This change adds an overload that accepts `ICollection<T>`. This is a safe addition as `ICollection<T>` derives from `IEnumerable<T>`, so there can be no ambiguity in overload resolution when passing a type that implements both interfaces.

Unfortunately the same is not true for other parameter types we may wish to statically assign to overloads such as `IReadOnlyCollection<T>` or `ICollection`. We can add runtime checks for these however.